Backend, GridView: fixed datatype for WriteLayer (int -> long) Wrote first test for GridView

+++ backend/Omnikeeper.Base/Entity/GridView/GridViewConfiguration.cs
@@ -5,7 +5,7 @@ namespace Omnikeeper.Base.Entity.GridView
     public class GridViewConfiguration
     {
         public bool ShowCIIDColumn { get; set; }
-        public int WriteLayer { get; set; }
+        public long WriteLayer { get; set; }
         public List<long> ReadLayerset { get; set; }
         public List<GridViewColumn> Columns { get; set; }
         public string Trait { get; set; }
@@ -15,6 +15,6 @@ namespace Omnikeeper.Base.Entity.GridView
     {
         public string SourceAttributeName { get; set; }
         public string ColumnDescription { get; set; }
-        public int? WriteLayer { get; set; }
+        public long? WriteLayer { get; set; }
     }
 }
